cleaned up, btb_project.py is now the main file
